msm: acpuclock-8960: Confirm power-collapse before disabling CPU clocks
Hotplug may not always be configured to involve power-collapse of
the CPU. When it is not, the CPU must remain clocked. If an HFPLL
is still selected as the clock source of the removed CPU, do not
disable it as part of the CPU_DEAD hotplug notification unless
msm_pm_verify_cpu_pc() confirms that the CPU has been collapsed.

This code can be significantly simplified if support for Krait v1
processors is ever removed from the driver. Krait v2 and later
allow changing the CPU clock source from another CPU even after
the affected CPU is collapsed. This feature could be used to change
the collapsed CPU's source away from the HFPLLs to something that
is always on (ex. QSB).

Signed-off-by: Matt Wagantall <mattw@codeaurora.org>
(cherry picked from commit 34c2d9628cb0fd471f43f92f04a6a9dc8df33129)

Conflicts:

	arch/arm/mach-msm/acpuclock-8960.c

Change-Id: I1ee2684359591000e6c756e497d14c8e2074916e
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>