radio: radio-tavarua: Re-org code for workqueue
When fm task is killed from task manager rpc failure is seen.
<3>[   72.817162] pmapp_rpc_req_reply: rpc failed! len = -512
<3>[   72.817173] fm_radio_shutdown: voting off failed with :(-512)
RPC failure is because data to be sent from apps, was on behalf of the
current process, but as the process is killed the current->pid given to
rpc is a non-existing pid.
To resolve the issue a valid pid is needed, which can be done by
queueing a work in existing workqueue and wait for completion.

CRs-fixed: 338027
Signed-off-by: Taniya Das <tdas@codeaurora.org>
(cherry picked from commit 9ac855f01e3898663dba342cddc2b3d39a4457dd)

Change-Id: I5b01b0dc08f4d4716460bdbda5a6176c70ec7c50
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>