Input: atmel_mxt_ts: Update config to newer version only
Add version checking so that config update only occurs when the config
data supplied in the platform data is of newer version. If the
controller is previously programmed with wrong version of config data,
based on controller type and firmware version, the driver will program
the latest config data present in the platform data to the controller.

CRs-fixed: 367335
Signed-off-by: Jing Lin <jinglin@codeaurora.org>
(cherry picked from commit 8fadad7bf01f18b1456eeb890c014c131026375c)

Change-Id: I02e36e571a8a892f176b66d625d2612be176dca6
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>