msm: bam_dmux: send powerdown ack to A2
When bam_dmux disconnects to the bam hardware in response to the A2
powering down, send an ack back to the A2 signaling that bam_dmux got
the powerdown signal and processed it.

This change optimizes the response time for back-to-back shutdown and
powerup sequences instead of requiring a hard coded delay in the A2 state
machine.

(cherry picked from commit 0b13a3522c4a68cb337398bb7bf66d3623d5f098)

Change-Id: I6e0d8036d189efba87873092562b8c5563c0db80
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>