power: pm8921-charger: Correct USB online reporting
When a USB charger is first connected to the device
the USB driver determines the maximum available charge
current. If the USB driver determines that a wall charger
is connected there is a possiblity of the USB power
supply type to falsely be reported as online.

Fix this by only reporting the USB online state if the
targeted maximum USB current is less than 500 mA. This
means that there was no wall charger ever connected and
AICL will not be utilized.

CRs-Fixed: 354361
Signed-off-by: David Keitel <dkeitel@codeaurora.org>
(cherry picked from commit af51571d389c397374e22584fdcb400a742d0dd6)

Change-Id: Ie8ada5dbdc60c8389792d5e677b31a5a9a41f193
Signed-off-by: Ram Kumar Chakravarthy Chebathini <rcheba@codeaurora.org>