Revert "radio-iris: Configure search direction as passed by FM APP"
As per V4L2 framework APP should pass the parameter as '1' to
search in forward direction and '0' to search in downward
direction. Reverting the change that was uploaded as it configured
the search direction directly that it received it from the APP.
APP should pass '1' for SEEK UP and '0' for SEEK DOWN . The driver
should check for the direction before configuring. To the FM SOC,
'0' means SEEK-UP and '1' means SEEK-DOWN. If APP passes '1' driver
understands that user is requesting SEEK-UP operation and will
configure the direction as '0'.

This reverts commit 4667d9d437c2fe5501db1fd741b4643452eda61b

(cherry picked from commit d7b5a77f8183754fb9391f47739d8a529dca4cd3)

Change-Id: I96827ebc09913bc1aeaae479f76dc19d7f0b0336
Signed-off-by: Anantha Krishnan <ananthk@codeaurora.org>