msm: watchdog: always manage interrupt on core 0
Previously the interrupt could end up being enabled or disabled on
either core. Since we switched to the percpu irq framework, it is a
problem if we enable the irq on one core and disable it on the other.
So, guarantee that the irq is always setup and removed on core 0.

CRs-fixed: 367280
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>
(cherry picked from commit b1e211ed27c6e09963513cc702f8682fa5fbbe4d)

Change-Id: Ib69c5808bca89453f3a222bed8eb66e4db4111fb
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>