USB: OTG: Fix PHY stuck issue in USB HOST mode
Disable PHY VBUS comparator if otg driver depends on PMIC
for VBUS and ID notifications. If enabled, VBUS comparator is
creating some noise to SESSVLD comparator and causing SESSVLD
comparator to trigger low due to which link state machine becomes
stuck as it operates solely on that value.

CRs-Fixed: 355991
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>

Conflicts:

	drivers/usb/otg/msm_otg.c

Change-Id: Idadc17410eb069c7d61ef2326002c57f578f5238
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>