msm: vidc: Avoid reconfig if EoS is received after SPS/PPS
After core init, there is a hang if the data buffers
are not sent to the core as the core is not responding
to the STOP command. With this commit, reconfig is
avoided and EoS done message is sent to the client.

CRs-Fixed: 360438
Signed-off-by: Saikumar Kondaparthi <saikumar@codeaurora.org>
(cherry picked from commit fcd8763cbf25679c999e363ac63e65a234e4192d)

Change-Id: Ib7a438742de113a97c6317a7420fdee15406bd4a
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>