ASoC: wcd9310/9304: Pull down Mic Bias output by default.
If Mic Bias is high Z by default and when the LDOH is enabled,
the Mic Bias output can pull-up to a non-zero voltage if there
is no loading or if the load leakage is very small.  Pulling
down the Mic Bias output so it is not in high Z to eliminate
the floating voltage.

CRs-Fixed: 368898
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>
(cherry picked from commit ccb29c6c7e8779ac6574f581d1f930a9a4cf6eb5)

Conflicts:

	sound/soc/codecs/wcd9304.c

Change-Id: I93ab762a9a8b03f003cc23b993fd54dee9b9b886
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>