msm_fb: hdmi: Turn on switch dev when connect to HDMI
Audio is not routed to TV after deep sleep and resume.
Userspace is dependent on switch device to indentify audio
route. When HDMI is turned on, this switch needs to be turned
on as well.

CRs-Fixed: 377159 368865
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>
(cherry picked from commit fd142681ad79a1b0e8ff007f146761ef8142e4c4)

Change-Id: Ibc717ceecba6ddc9191ad5b22a8b09900281f4d6
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>