msm: clock-8960: Move DSI resets functionality to their own clocks
The DSI core resets are currently associated with the DSI byte clocks
because the DSI core clocks are managed by the DSI driver directly,
not the clock driver. Move the reset functionality out of the byte
and into their own reset-only branch clocks so that the footswitch
driver, which needs to assert these resets, can do so without
involving the unrelated byte clock.

Change-Id: I65a84e97181c5461c5f2e0b0af19dc6b0b1d2557
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>