msm: footswitch-8x60: Reset DSI domain along with MDP collapse/restore
Some DSI-related logic lives in MDP's collapsible GDHS domain for
896 and 8064. Assert the reset to this domain when the reset of MDP
is reset. Use of the reset-only clocks added here assumes the
display driver is making sure the DSI clocks (not managed by the
clock driver) are enabled for the duration of the collapse and
restore sequences so that the resets propagate.

Change-Id: Ife1b13e22e35b322ff5da75280b3c8d9d5027743
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>