power: pm8921-charger: Control CHG_USB_SUSPEND enabling boost
Toggle the CHG_USB_SUSPEND bit when utilizing the VBUS boost.
This bit controls ensures that no current is being drawn from
a USB device and the charger runs off the battery.

This makes sure that when pm8921_disable_source_current is called
the correct setting is written to the suspend bit depending
on the disable flag.

CRs-Fixed: 393498
Change-Id: I2acfae5aec19165a2196c50c31569c999cb090db
Signed-off-by: David Keitel <dkeitel@codeaurora.org>